backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
sqlalchemy
psycopg2-binary
supabase
pdfplumber
python-docx
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pandas
openpyxl
fpdf2
pytest
httpx
pytest-asyncio
